refactoring: Header 컴포넌트 및 하위 컴포넌트
Header 폴더로 이동
useStyle를 컴포넌트 외부로 이동
avata -> avatar 로 이름 변경
